Move keybindings to separate module
